(#1859) - WIP - Performance tests for map/reduce
So, the good news is that persistent mapreduce
is faster than the old temporary in-memory view.

The bad news is that it's not *much* faster,
only about 1.5 or twice as faster, rather than
the order of magnitude difference we were expecting.
That's even with discounting the initial buildup
with the first query.

Luckily it seems there are some low-hanging fruit
we can grab to make it faster